Extract `throwSyntaxError`
